+++ packages/media/media-viewer/src/newgen/styled.tsx
@@ -6,15 +6,18 @@ import styled, { StyledComponentClass } from 'styled-components';
 import { MediaType } from '@atlaskit/media-core';
 // @ts-ignore: unused variable
 // prettier-ignore
-import { HTMLAttributes, VideoHTMLAttributes, ImgHTMLAttributes, ComponentClass, ClassAttributes } from 'react';
+import { HTMLAttributes, VideoHTMLAttributes, AudioHTMLAttributes, ImgHTMLAttributes, ComponentClass, ClassAttributes } from 'react';
 import {
   akColorY200,
   akColorP200,
   akColorB300,
   akColorN400,
+  akBorderRadius,
 } from '@atlaskit/util-shared-styles';
 
 const overlayZindex = 999;
+
+export const hideControlsClassName = 'mvng-hide-controls';
 
 export const Blanket = styled.div`
   position: fixed;
@@ -66,6 +69,17 @@ export const CloseButtonWrapper = styled.div`
   z-index: ${overlayZindex + 2};
 `;
 
+export const ZoomWrapper = styled.div`
+  position: absolute;
+  bottom: 10px;
+  left: 50%;
+  transform: translateX(-50%);
+
+  button {
+    margin-right: 10px;
+  }
+`;
+
 const handleControlsVisibility = ({ showControls }: ContentWrapperProps) => `
   transition: opacity .3s;
   opacity: ${showControls ? '1' : '0'};
@@ -79,15 +93,11 @@ export const ContentWrapper = styled.div`
   align-items: center;
   justify-content: center;
 
-  ${HeaderWrapper} {
+  .${hideControlsClassName} {
     ${handleControlsVisibility};
   }
 
-  ${ArrowsWrapper} {
-    ${handleControlsVisibility};
-  }
-
-  ${CloseButtonWrapper} {
+  ${ZoomWrapper} {
     ${handleControlsVisibility};
   }
 `;
@@ -99,7 +109,8 @@ export const ErrorMessage = styled.div`
 `;
 
 export const Img: ComponentClass<ImgHTMLAttributes<{}>> = styled.img`
-  max-width: 100%;
+  transition: transform 0.2s;
+  transform-origin: center;
 `;
 
 export const Video: ComponentClass<VideoHTMLAttributes<{}>> = styled.video`
@@ -118,11 +129,11 @@ export const Arrow = styled.span`
   cursor: pointer;
 `;
 
-export const LeftWrapper = ArrowWrapper.extend`
+export const LeftWrapper = styled(ArrowWrapper)`
   text-align: left;
 `;
 
-export const RightWrapper = ArrowWrapper.extend`
+export const RightWrapper = styled(ArrowWrapper)`
   text-align: right;
 `;
 
@@ -135,6 +146,14 @@ export const LeftHeader = styled.div`
   flex: 0.8;
 `;
 
+export const ImageWrapper = styled.div`
+  width: 100vw;
+  height: 100vh;
+  overflow: auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
 export const MetadataWrapper = styled.div`
   display: flex;
 `;
@@ -154,7 +173,7 @@ export const MetadataIconWrapper = styled.div`
   padding-right: 12px;
 `;
 
-const colors = {
+export const colors = {
   image: akColorY200,
   audio: akColorP200,
   video: '#ff7143',
@@ -177,4 +196,42 @@ export const RightHeader = styled.div`
   flex: 0.2;
   text-align: right;
   margin-right: 50px;
+`;
+
+export const AudioPlayer = styled.div`
+  border-radius: ${akBorderRadius};
+  align-items: center;
+  justify-content: center;
+  width: 400px;
+  height: 250px;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+`;
+
+export const Audio = styled.audio`
+  width: 100%;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+`;
+
+export const AudioCover = styled.img`
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+`;
+
+export const DefaultCoverWrapper = styled.div`
+  background-color: ${colors.audio};
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  > * {
+    transform: scale(2);
+  }
 `;