+++ packages/core/pagination/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/pagination
 
+## 5.0.1
+- [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
+  - @atlaskit/button@7.2.5
+  - @atlaskit/docs@3.0.4
+
 ## 5.0.0
 - [major] Bump to React 16.3. [4251858](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4251858)