+++ packages/css-packs/reduced-ui-pack/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/reduced-ui-pack
 
+## 8.9.1
+- [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
+  - @atlaskit/icon@11.3.0
+  - @atlaskit/docs@3.0.4
+  - @atlaskit/css-reset@2.0.2
+
 ## 8.9.0
 - [minor] Add divider from editor  [5cbb8a6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5cbb8a6)
 - [minor] Add divider fabric icon [8b794ed](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8b794ed)